--- src/packages/core/repository/copy/copy-data-source.interface.ts
@@ -1,5 +1,0 @@
-import type { UmbDataSourceResponse } from '@umbraco-cms/backoffice/repository';
-
-export interface UmbCopyDataSource {
-	copy(unique: string, targetUnique: string | null): Promise<UmbDataSourceResponse<string>>;
-}
