--- src/packages/core/repository/copy/copy-repository.interface.ts
@@ -1,5 +1,0 @@
-import type { UmbRepositoryResponse } from '../types.js';
-
-export interface UmbCopyRepository {
-	copy(unique: string, targetUnique: string): Promise<UmbRepositoryResponse<string>>;
-}
