--- src/packages/core/repository/copy/index.ts
@@ -1,2 +1,0 @@
-export type { UmbCopyDataSource } from './copy-data-source.interface.js';
-export type { UmbCopyRepository } from './copy-repository.interface.js';
