+++ src/packages/core/repository/duplicate/duplicate-data-source.interface.ts
@@ -1,0 +1,5 @@
+import type { UmbDataSourceResponse } from '@umbraco-cms/backoffice/repository';
+
+export interface UmbDuplicateDataSource {
+	duplicate(unique: string, targetUnique: string | null): Promise<UmbDataSourceResponse<string>>;
+}
