+++ src/packages/core/repository/duplicate/duplicate-repository.interface.ts
@@ -1,0 +1,5 @@
+import type { UmbRepositoryResponse } from '../types.js';
+
+export interface UmbDuplicateRepository {
+	duplicate(unique: string, targetUnique: string): Promise<UmbRepositoryResponse<string>>;
+}
