+++ src/packages/core/repository/duplicate/index.ts
@@ -1,0 +1,2 @@
+export type { UmbDuplicateDataSource } from './duplicate-data-source.interface.js';
+export type { UmbDuplicateRepository } from './duplicate-repository.interface.js';
