+++ src/packages/core/repository/index.ts
@@ -6,5 +6,5 @@ export * from './detail/index.js';
 
 export type { UmbDataSourceResponse, UmbDataSourceErrorResponse } from './data-source-response.interface.js';
 export type { UmbMoveDataSource, UmbMoveRepository } from './move/index.js';
-export type { UmbCopyDataSource, UmbCopyRepository } from './copy/index.js';
+export type { UmbDuplicateDataSource, UmbDuplicateRepository } from './duplicate/index.js';
 export type { UmbPagedModel } from './types.js';
