+++ src/packages/core/tree/reload-tree-item-children/manifests.ts
@@ -1,3 +1,3 @@
-import { manifest as reloadTreeItemChildren } from './reload-tree-item-children.action.kind.js';
+import { manifest as reloadTreeItemChildrenKind } from './reload-tree-item-children.action.kind.js';
 
-export const manifests = [reloadTreeItemChildren];
+export const manifests = [reloadTreeItemChildrenKind];
