--- src/packages/data-type/repository/copy/index.ts
@@ -1,2 +1,0 @@
-export { UmbCopyDataTypeRepository } from './data-type-copy.repository.js';
-export { UMB_COPY_DATA_TYPE_REPOSITORY_ALIAS as COPY_DATA_TYPE_REPOSITORY_ALIAS } from './manifests.js';
