+++ src/packages/data-type/repository/duplicate/index.ts
@@ -1,0 +1,2 @@
+export { UmbCopyDataTypeRepository } from './data-type-copy.repository.js';
+export { UMB_DUPLICATE_DATA_TYPE_REPOSITORY_ALIAS } from './manifests.js';
