+++ src/packages/data-type/repository/manifests.ts
@@ -1,4 +1,4 @@
-import { manifests as copyManifests } from './copy/manifests.js';
+import { manifests as copyManifests } from './duplicate/manifests.js';
 import { manifests as detailManifests } from './detail/manifests.js';
 import { manifests as itemManifests } from './item/manifests.js';
 import { manifests as moveManifests } from './move/manifests.js';
