+++ src/packages/documents/documents/entity-bulk-actions/duplicate/duplicate.action.ts
@@ -1,6 +1,6 @@
 import { UmbEntityBulkActionBase } from '@umbraco-cms/backoffice/entity-bulk-action';
 
-export class UmbDocumentCopyEntityBulkAction extends UmbEntityBulkActionBase<object> {
+export class UmbDocumentDuplicateEntityBulkAction extends UmbEntityBulkActionBase<object> {
 	async execute() {
 		console.log('execute bulk copy');
 	}
